+++ app/src/main/java/com/example/packlest/PackingListActivity.java
@@ -8,12 +8,13 @@ import android.view.MenuItem;
 import android.widget.ListView;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.UUID;
 
 public class PackingListActivity extends AppCompatActivity {
-    private ListView itemListView;
+    private ListView listView;
     private ListViewItemCheckboxAdapter dataAdapter;
     private PackingList filteredPackingList;
     private static final String TAG = "PackingListActivity";
@@ -28,9 +29,9 @@ public class PackingListActivity extends AppCompatActivity {
         filteredPackingList = new PackingList(PacklestApplication.getInstance().packlestData.packingLists.get(packingListUuid));
         setTitle(filteredPackingList.name);
 
-        itemListView = findViewById(R.id.listViewItems);
+        listView = findViewById(R.id.listViewItems);
         dataAdapter = new ListViewItemCheckboxAdapter(this, filteredPackingList);
-        itemListView.setAdapter(dataAdapter);
+        listView.setAdapter(dataAdapter);
 
         setListViewOnItemClickListener();
     }
@@ -54,19 +55,17 @@ public class PackingListActivity extends AppCompatActivity {
                 if (dataAdapter.filter_state == FILTER_STATE.NONE) {
                     Log.v(TAG, "Filtering out un-added itemInstances");
                     menuItem.setIcon(R.drawable.ic_filter);
-                    dataAdapter.getFilter().filter(FILTER_STATE.ADDED_ONLY.name());
                     dataAdapter.filter_state = FILTER_STATE.ADDED_ONLY;
                 } else if (dataAdapter.filter_state == FILTER_STATE.ADDED_ONLY) {
                     Log.v(TAG, "Filtering out checked-itemInstances too");
-                    dataAdapter.getFilter().filter(FILTER_STATE.UNCHECKED_ONLY.name());
                     dataAdapter.filter_state = FILTER_STATE.UNCHECKED_ONLY;
                     menuItem.setIcon(R.drawable.ic_filter_remove);
                 } else if (dataAdapter.filter_state == FILTER_STATE.UNCHECKED_ONLY){
                     Log.v(TAG, "Resetting the filter");
-                    dataAdapter.getFilter().filter(FILTER_STATE.NONE.name());
                     dataAdapter.filter_state = FILTER_STATE.NONE;
                     menuItem.setIcon(R.drawable.ic_filter_outline);
                 }
+                dataAdapter.getFilter().filter(dataAdapter.filter_state.name());
                 dataAdapter.notifyDataSetChanged();
                 break;
             case R.id.edit_packing_list:
@@ -77,8 +76,15 @@ public class PackingListActivity extends AppCompatActivity {
                 break;
             case R.id.delete_packing_list:
                 Log.v(TAG, "Deleting packing list");
-                PacklestApplication.getInstance().packlestData.deletePackingList(filteredPackingList.uuid);
-                finish();
+                new AlertDialog.Builder(this)
+                        .setTitle("Confirm Deletion")
+                        .setMessage("Do you really want to delete: " + filteredPackingList.name + "?")
+                        .setIcon(android.R.drawable.ic_menu_delete)
+                        .setPositiveButton(android.R.string.yes, (dialog, button) -> {
+                            PacklestApplication.getInstance().packlestData.deletePackingList(filteredPackingList.uuid);
+                            finish();
+                        })
+                        .setNegativeButton(android.R.string.cancel, null).show();
                 break;
             case R.id.un_add_all_items:
                 Log.v(TAG, "Un-adding all itemInstances");
@@ -104,7 +110,7 @@ public class PackingListActivity extends AppCompatActivity {
 
     private void setListViewOnItemClickListener() {
         Log.v(TAG, "Item Clicked");
-        itemListView.setOnItemClickListener((parent, view, position, id) -> {
+        listView.setOnItemClickListener((parent, view, position, id) -> {
             ItemInstance itemInstance = dataAdapter.getItem(position);
             Item item = PacklestApplication.getInstance().packlestData.items.get(itemInstance.itemUuid);
 
@@ -113,12 +119,6 @@ public class PackingListActivity extends AppCompatActivity {
             intent.putExtra("packingListUuid", filteredPackingList.uuid);
             startActivityForResult(intent, PacklestApplication.IGNORED_REQUEST_CODE);
         });
-    }
-
-    @Override
-    public void onBackPressed() {
-        Log.v(TAG, "Back Pressed");
-        finish();
     }
 
     @Override
