+++ app/src/main/java/com/example/packlest/TripParameterEditorActivity.java
@@ -5,6 +5,7 @@ import android.util.Log;
 import android.view.MenuItem;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 
 import java.util.UUID;
 
@@ -36,8 +37,15 @@ public class TripParameterEditorActivity extends AbstractEditorActivity {
     public boolean onOptionsItemSelected(@NonNull MenuItem menuItem) {
         if (menuItem.getItemId() == R.id.delete_item) {
             Log.v(TAG, "Deleting Trip Parameter");
-            PacklestApplication.getInstance().packlestData.deleteTripParameter(tripParameter.uuid);
-            finish();
+            new AlertDialog.Builder(this)
+                    .setTitle("Confirm Deletion")
+                    .setMessage("Do you really want to delete: " + tripParameter.name + "?")
+                    .setIcon(android.R.drawable.ic_menu_delete)
+                    .setPositiveButton(android.R.string.yes, (dialog, button) -> {
+                        PacklestApplication.getInstance().packlestData.deleteTripParameter(tripParameter.uuid);
+                        finish();
+                    })
+                    .setNegativeButton(android.R.string.cancel, null).show();
         }
 
         return super.onOptionsItemSelected(menuItem);
